+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fixed `MutableStateFlow.compareAndSet` for `.toMutable`/`.beforeUpdate`/`.afterUpdate`.
 * Added `Flow.stateOnDemand` variant with `initial` value.
 * `CoroutineLauncher.launch` now synchronously increments the loading counter to avoid UI flickering issues in edge cases.
+* Added `MutableStateFlow<T>.collectAsMutableStateWithLifecycle()`.
 * Added `childReactiveState` variant which takes an arbitrary event handler, so the parent ReactiveState doesn't have to implement the whole events interface.
 * Removed direct dependency on JUnit 4, so you can choose more freely which JUnit version to use.